examples: simplfied the low_consumption example into a single file.
