+++ examples/low_consumption/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS rclc_int32_publisher.c main.c 
+idf_component_register(SRCS main.c 
                     INCLUDE_DIRS ""
                     )
