+++ examples/low_consumption/main/main.c
@@ -1,140 +1,114 @@
-#include "uxr/client/config.h"
+#include <string.h>
+#include <stdio.h>
+#include <unistd.h>
 
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
-
-#include <driver/uart.h>
-#include <driver/gpio.h>
-
-#include <string.h>
-#include "freertos/event_groups.h"
-#include "esp_system.h"
-#include "esp_wifi.h"
-#include "esp_event.h"
 #include "esp_log.h"
 #include "esp_system.h"
-#include "nvs_flash.h"
+
+#include <uros_network_interfaces.h>
+#include <rcl/rcl.h>
+#include <rcl/error_handling.h>
+#include <std_msgs/msg/int32.h>
+#include <rclc/rclc.h>
+#include <rclc/executor.h>
+#include <rmw_uros/options.h>
+#include "uxr/client/config.h"
+
 #ifdef CONFIG_PM_ENABLE
 #include "esp_pm.h"
 #endif /* CONFIG_PM_ENABLE */
 
-#include "lwip/err.h"
-#include "lwip/sys.h"
+#define RCCHECK(fn) { rcl_ret_t temp_rc = fn; if((temp_rc != RCL_RET_OK)){printf("Failed status on line %d: %d. Aborting.\n",__LINE__,(int)temp_rc);vTaskDelete(NULL);}}
+#define RCSOFTCHECK(fn) { rcl_ret_t temp_rc = fn; if((temp_rc != RCL_RET_OK)){printf("Failed status on line %d: %d. Continuing.\n",__LINE__,(int)temp_rc);}}
 
-void appMain(void *argument);
+rcl_publisher_t publisher;
+std_msgs__msg__Int32 msg;
 
-#define ESP_WIFI_SSID      CONFIG_ESP_WIFI_SSID
-#define ESP_WIFI_PASS      CONFIG_ESP_WIFI_PASSWORD
-#define ESP_MAXIMUM_RETRY  CONFIG_ESP_MAXIMUM_RETRY
+#ifdef CONFIG_PM_ENABLE
+esp_pm_lock_handle_t pmlock;
+#endif /* CONFIG_PM_ENABLE */
 
-static EventGroupHandle_t s_wifi_event_group;
-
-#define WIFI_CONNECTED_BIT BIT0
-#define WIFI_FAIL_BIT      BIT1
-
-static const char *TAG = "wifi station";
-static int s_retry_num = 0;
-
-static void event_handler(void* arg, esp_event_base_t event_base,
-                                int32_t event_id, void* event_data)
+void timer_callback(rcl_timer_t * timer, int64_t last_call_time)
 {
-    if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_START) {
-        esp_wifi_connect();
-    } else if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_DISCONNECTED) {
-        if (s_retry_num < ESP_MAXIMUM_RETRY) {
-            esp_wifi_connect();
-            s_retry_num++;
-            ESP_LOGI(TAG, "retry to connect to the AP");
-        } else {
-            xEventGroupSetBits(s_wifi_event_group, WIFI_FAIL_BIT);
-        }
-        ESP_LOGI(TAG,"connect to the AP fail");
-    } else if (event_base == IP_EVENT && event_id == IP_EVENT_STA_GOT_IP) {
-        ip_event_got_ip_t* event = (ip_event_got_ip_t*) event_data;
-        ESP_LOGI(TAG, "got ip:" IPSTR, IP2STR(&event->ip_info.ip));
-        s_retry_num = 0;
-        xEventGroupSetBits(s_wifi_event_group, WIFI_CONNECTED_BIT);
-    }
+	RCLC_UNUSED(last_call_time);
+	if (timer != NULL) {
+#ifdef CONFIG_PM_ENABLE
+                esp_pm_lock_acquire(pmlock);	// disable wifi sleep mode
+#endif /* CONFIG_PM_ENABLE */
+		RCSOFTCHECK(rcl_publish(&publisher, &msg, NULL));
+#ifdef CONFIG_PM_ENABLE
+                esp_pm_lock_release(pmlock);	// allow wifi sleep mode 
+#endif /* CONFIG_PM_ENABLE */
+		msg.data++;
+	}
 }
 
-void wifi_init_sta()
+void micro_ros_task(void * arg)
 {
-    s_wifi_event_group = xEventGroupCreate();
-
-    ESP_ERROR_CHECK(esp_netif_init());
-
-    ESP_ERROR_CHECK(esp_event_loop_create_default());
-    esp_netif_create_default_wifi_sta();
-
-    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
-    ESP_ERROR_CHECK(esp_wifi_init(&cfg));
-
-    ESP_ERROR_CHECK(esp_event_handler_register(WIFI_EVENT, ESP_EVENT_ANY_ID, &event_handler, NULL));
-    ESP_ERROR_CHECK(esp_event_handler_register(IP_EVENT, IP_EVENT_STA_GOT_IP, &event_handler, NULL));
-
-    wifi_config_t wifi_config = {
-        .sta = {
-            .ssid = ESP_WIFI_SSID,
-            .password = ESP_WIFI_PASS,
-#ifdef CONFIG_PM_ENABLE
-            .listen_interval = 5,
-/* Listen interval for ESP32 station to receive beacon when WIFI_PS_MAX_MODEM is set. 
-   Units: AP beacon intervals. Defaults to 3 if set to 0. */
-#endif /* CONFIG_PM_ENABLE */
-        },
-    };
-    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_STA) );
-    ESP_ERROR_CHECK(esp_wifi_set_config(ESP_IF_WIFI_STA, &wifi_config) );
-    ESP_ERROR_CHECK(esp_wifi_start() );
-
-    ESP_LOGI(TAG, "wifi_init_sta finished.");
+	rcl_allocator_t allocator = rcl_get_default_allocator();
+	rclc_support_t support;
 
 #ifdef CONFIG_PM_ENABLE
-    ESP_LOGI(TAG, "esp_wifi_set_ps().");
-    esp_wifi_set_ps(WIFI_PS_MAX_MODEM);
-    /* Call esp_wifi_set_ps(WIFI_PS_MIN_MODEM) to enable Modem-sleep minimum power save mode 
-    or esp_wifi_set_ps(WIFI_PS_MAX_MODEM) to enable Modem-sleep maximum power save mode after 
-    calling esp_wifi_init(). When station connects to AP, Modem-sleep will start. 
-    When station disconnects from AP, Modem-sleep will stop. */
+    esp_pm_lock_create(ESP_PM_NO_LIGHT_SLEEP, 0, "pmlock", &pmlock);
 #endif /* CONFIG_PM_ENABLE */
 
-    /* Waiting until either the connection is established (WIFI_CONNECTED_BIT) or connection failed for the maximum
-     * number of re-tries (WIFI_FAIL_BIT). The bits are set by event_handler() (see above) */
-    EventBits_t bits = xEventGroupWaitBits(s_wifi_event_group,
-            WIFI_CONNECTED_BIT | WIFI_FAIL_BIT,
-            pdFALSE,
-            pdFALSE,
-            portMAX_DELAY);
+	rcl_init_options_t init_options = rcl_get_zero_initialized_init_options();
+	RCCHECK(rcl_init_options_init(&init_options, allocator));
+	rmw_init_options_t* rmw_options = rcl_init_options_get_rmw_init_options(&init_options);
 
-    /* xEventGroupWaitBits() returns the bits before the call returned, hence we can test which event actually
-     * happened. */
-    if (bits & WIFI_CONNECTED_BIT) {
-        ESP_LOGI(TAG, "connected to ap SSID:%s", ESP_WIFI_SSID);
-    } else if (bits & WIFI_FAIL_BIT) {
-        ESP_LOGI(TAG, "Failed to connect to SSID:%s", ESP_WIFI_SSID);
-    } else {
-        ESP_LOGE(TAG, "UNEXPECTED EVENT");
-    }
+	// Static Agent IP and port can be used instead of autodisvery.
+	RCCHECK(rmw_uros_options_set_udp_address(CONFIG_MICRO_ROS_AGENT_IP, CONFIG_MICRO_ROS_AGENT_PORT, rmw_options));
+	//RCCHECK(rmw_uros_discover_agent(rmw_options));
 
-    ESP_ERROR_CHECK(esp_event_handler_unregister(IP_EVENT, IP_EVENT_STA_GOT_IP, &event_handler));
-    ESP_ERROR_CHECK(esp_event_handler_unregister(WIFI_EVENT, ESP_EVENT_ANY_ID, &event_handler));
-    vEventGroupDelete(s_wifi_event_group);
+	// create init_options
+	RCCHECK(rclc_support_init_with_options(&support, 0, NULL, &init_options, &allocator));
+
+	// create node
+	rcl_node_t node;
+	RCCHECK(rclc_node_init_default(&node, "esp32_int32_publisher", "", &support));
+
+	// create publisher
+	RCCHECK(rclc_publisher_init_default(
+		&publisher,
+		&node,
+		ROSIDL_GET_MSG_TYPE_SUPPORT(std_msgs, msg, Int32),
+		"freertos_int32_publisher"));
+
+	// create timer,
+	rcl_timer_t timer;
+	const unsigned int timer_timeout = 1000;
+	RCCHECK(rclc_timer_init_default(
+		&timer,
+		&support,
+		RCL_MS_TO_NS(timer_timeout),
+		timer_callback));
+
+	// create executor
+	rclc_executor_t executor;
+	RCCHECK(rclc_executor_init(&executor, &support.context, 1, &allocator));
+	RCCHECK(rclc_executor_add_timer(&executor, &timer));
+
+	msg.data = 0;
+
+	while(1){
+		rclc_executor_spin_some(&executor, RCL_MS_TO_NS(100));
+		usleep(10000);
+	}
+
+	// free resources
+	RCCHECK(rcl_publisher_fini(&publisher, &node));
+	RCCHECK(rcl_node_fini(&node));
+
+  	vTaskDelete(NULL);
 }
 
 void app_main(void)
 {   
-    // Start networkign if required
 #ifdef UCLIENT_PROFILE_UDP
-    //Initialize NVS
-    esp_err_t ret = nvs_flash_init();
-    if (ret == ESP_ERR_NVS_NO_FREE_PAGES || ret == ESP_ERR_NVS_NEW_VERSION_FOUND) {
-      ESP_ERROR_CHECK(nvs_flash_erase());
-      ret = nvs_flash_init();
-    }
-    ESP_ERROR_CHECK(ret);
-
-    ESP_LOGI(TAG, "ESP_WIFI_MODE_STA");
-    wifi_init_sta();
+    // Start the networking if required
+    ESP_ERROR_CHECK(uros_network_interface_initialize());
 #endif  // UCLIENT_PROFILE_UDP
 
 #ifdef CONFIG_PM_ENABLE
@@ -158,6 +132,11 @@ void app_main(void)
     ESP_ERROR_CHECK(esp_pm_configure(&pm_config));
 #endif /* CONFIG_PM_ENABLE */
 
-    // start microROS task
-    xTaskCreate(appMain, "uros_task", CONFIG_MICRO_ROS_APP_STACK, NULL, 5, NULL);
+    //pin micro-ros task in APP_CPU to make PRO_CPU to deal with wifi:
+    xTaskCreate(micro_ros_task, 
+            "uros_task", 
+            CONFIG_MICRO_ROS_APP_STACK, 
+            NULL,
+            CONFIG_MICRO_ROS_APP_TASK_PRIO, 
+            NULL);
 }
