+++ examples/low_consumption/sdkconfig.defaults
@@ -1,3 +1,6 @@
+CONFIG_PM_ENABLE=y
+CONFIG_PM_DFS_INIT_AUTO=y
+CONFIG_FREERTOS_USE_TICKLESS_IDLE=y
 CONFIG_ESP_MAIN_TASK_STACK_SIZE=3000
 CONFIG_ESP_TASK_WDT=n
